Fix scalar quantization example
* The search_params were being destroyed before use in cuvsCagraSearch with the quantized index
* The queries_d memory was being freed before use in the SQ transform
* The cuvsScalarQuantizerTransform function was failing - since both the input/output tensors need
to be both on host memory or both on device memory (and was being passed the input on host and the
output on device for the dataset transform).

Fix by moving cleanup functions to the end, and using the dataset on device for the SQ
transform.

(note: Since we're not checking the return codes on any functions, it wasn't obvious that the
cuvsScalarQuantizerTransform function failed . This problem goes beyond this example, and
I will fix in a separate PR. )